refactored frontend so that the Pages are in different directories
