+++ packages/react-frontend/src/pages/Com_Home_post_pages/LandingPage/Posts.jsx
@@ -1,12 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import Newpostpopup from './Newpostpopup.jsx';
 import './Post.css';
 import { NavLink, redirect, useNavigate, useParams } from 'react-router-dom';
 import { white } from 'color-name';
 
 function TableBody(props) {
   const nav = useNavigate();
-
   function getCommunityName(post) {
     return props.getCommunityByPostid(post.id);
   }
