+++ packages/react-frontend/src/pages/Com_Home_post_pages/NotFound.jsx
@@ -1,3 +1,9 @@
-export default function NotFound() {
-  return <h1>Hello! The page does not exist!</h1>;
+export default function NotFound(props) {
+  if(props.communityName && props.postHeader){
+    return <h1>Hello! Post {props.communityName}/{props.postHeader} does not exist</h1>;
+  }else if (props.communityName){
+    return <h1>Hello! Community {props.communityName} does not exist</h1>;
+  } else {
+  return <h1>404 Not Found</h1>;
+  }
 }
